Crate sp_consensus_subspace

Primitives for Subspace consensus.

Modules

consensus

Consensus-related runtime interface

digests

Private implementation details of Subspace consensus digests.

inherents

Inherents for Subspace consensus

Structs

KzgExtension

A KZG extension.

PosExtension

A Poof of space extension.

PotExtension

A Poof of time extension.

PotNextSlotInput

Next slot input for proof of time evaluation

PotParametersChange

Change of parameters to apply to PoT chain

SignedVote

Signed farmer vote.

SolutionRanges

Subspace solution ranges used for challenges.

WrappedPotOutput

Wrapped proof of time output for the purposes of runtime interface.

WrappedSolution

Wrapped solution for the purposes of runtime interface.

WrappedVerifySolutionParams

Wrapped solution verification parameters for the purposes of runtime interface.

Enums

ChainConstants

Subspace blockchain constants.

PotParameters

Proof of time parameters

SubspaceJustification

Subspace justification

Vote

Farmer vote.

Traits

SubspaceApi

API necessary for block authorship with Subspace.