sp_consensus_subspace
Module digests

Private implementation details of Subspace consensus digests.

Structs

DeriveNextSolutionRangeParams

Params used to derive the next solution range.

NextDigestsVerificationParams

Type that holds the parameters to derive and verify next digest items.

SubspaceDigestItems

Digest items extracted from a header into convenient form

Enums

Error

Digest error

ErrorDigestType

Various kinds of digest types used in errors

PreDigest

A Subspace pre-runtime digest. This contains all data required to validate a block and for the Subspace runtime module.

PreDigestPotInfo

Proof of time information in pre-digest

Traits

CompatibleDigestItem

A digest item which is usable with Subspace consensus.

Functions

derive_next_solution_range

Derives next solution range if era duration interval has met.

extract_pre_digest

Extract the Subspace pre digest from the given header. Pre-runtime digests are mandatory, the function will return Err if none is found.

extract_subspace_digest_items

Extract the Subspace global randomness from the given header.

verify_next_digests

Derives and verifies next digest items based on their respective intervals.