Function sp_consensus_subspace::check_header

pub fn check_header<Header, RewardAddress>(
    params: VerificationParams<'_, Header>,
    pre_digest: Option<PreDigest<FarmerPublicKey, RewardAddress>>
) -> Result<CheckedHeader<Header, VerifiedHeaderInfo<RewardAddress>>, VerificationError<Header>> where
    Header: HeaderT,
    RewardAddress: Decode, 

Check a header has been signed correctly and whether solution is correct. If the slot is too far in the future, an error will be returned. If successful, returns the pre-header and the digest item containing the seal.

The seal must be the last digest. Otherwise, the whole header is considered unsigned. This is required for security and must not be changed.

This digest item will always return Some when used with as_subspace_pre_digest.

pre_digest argument is optional in case it is available to avoid doing the work of extracting it from the header twice.