Trait sp_consensus_subspace::SubspaceApi

pub trait SubspaceApi<Block: BlockT, RewardAddress: Encode + Decode>: Core<Block> {

    fn slot_duration(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Duration, ApiError> { ... }
    fn slot_duration_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<Duration, ApiError> { ... }
    fn global_randomnesses(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<GlobalRandomnesses, ApiError> { ... }
    fn global_randomnesses_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<GlobalRandomnesses, ApiError> { ... }
    fn solution_ranges(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<SolutionRanges, ApiError> { ... }
    fn solution_ranges_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<SolutionRanges, ApiError> { ... }
    fn submit_report_equivocation_extrinsic(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        equivocation_proof: EquivocationProof<Block::Header>
    ) -> Result<Option<()>, ApiError> { ... }
    fn submit_report_equivocation_extrinsic_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        equivocation_proof: EquivocationProof<Block::Header>
    ) -> Result<Option<()>, ApiError> { ... }
    fn submit_vote_extrinsic(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
    ) -> Result<(), ApiError> { ... }
    fn submit_vote_extrinsic_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
    ) -> Result<(), ApiError> { ... }
    fn is_in_block_list(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        farmer_public_key: &FarmerPublicKey
    ) -> Result<bool, ApiError> { ... }
    fn is_in_block_list_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        farmer_public_key: &FarmerPublicKey
    ) -> Result<bool, ApiError> { ... }
    fn total_pieces(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<NonZeroU64, ApiError> { ... }
    fn total_pieces_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<NonZeroU64, ApiError> { ... }
    fn records_root(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        segment_index: SegmentIndex
    ) -> Result<Option<RecordsRoot>, ApiError> { ... }
    fn records_root_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        segment_index: SegmentIndex
    ) -> Result<Option<RecordsRoot>, ApiError> { ... }
    fn extract_root_blocks(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        ext: &Block::Extrinsic
    ) -> Result<Option<Vec<RootBlock>>, ApiError> { ... }
    fn extract_root_blocks_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        ext: &Block::Extrinsic
    ) -> Result<Option<Vec<RootBlock>>, ApiError> { ... }
    fn root_plot_public_key(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<FarmerPublicKey>, ApiError> { ... }
    fn root_plot_public_key_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<Option<FarmerPublicKey>, ApiError> { ... }
    fn should_adjust_solution_range(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<bool, ApiError> { ... }
    fn should_adjust_solution_range_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<bool, ApiError> { ... }
    fn chain_constants(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<ChainConstants, ApiError> { ... }
    fn chain_constants_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<ChainConstants, ApiError> { ... }

}

API necessary for block authorship with Subspace.

Provided Methods

fn slot_duration(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Duration, ApiError>

The slot duration in milliseconds for Subspace.

fn slot_duration_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Duration, ApiError>

The slot duration in milliseconds for Subspace.

fn global_randomnesses(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<GlobalRandomnesses, ApiError>

Global randomnesses used for deriving global challenges.

fn global_randomnesses_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<GlobalRandomnesses, ApiError>

Global randomnesses used for deriving global challenges.

fn solution_ranges(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<SolutionRanges, ApiError>

Solution ranges.

fn solution_ranges_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<SolutionRanges, ApiError>

Solution ranges.

fn submit_report_equivocation_extrinsic(
&self,
__runtime_api_at_param__: &BlockId<Block>,
equivocation_proof: EquivocationProof<Block::Header>
) -> Result<Option<()>, ApiError>

Submits an unsigned extrinsic to report an equivocation. The caller must provide the equivocation proof. The extrinsic will be unsigned and should only be accepted for local authorship (not to be broadcast to the network). This method returns None when creation of the extrinsic fails, e.g. if equivocation reporting is disabled for the given runtime (i.e. this method is hardcoded to return None). Only useful in an offchain context.

fn submit_report_equivocation_extrinsic_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
equivocation_proof: EquivocationProof<Block::Header>
) -> Result<Option<()>, ApiError>

Submits an unsigned extrinsic to report an equivocation. The caller must provide the equivocation proof. The extrinsic will be unsigned and should only be accepted for local authorship (not to be broadcast to the network). This method returns None when creation of the extrinsic fails, e.g. if equivocation reporting is disabled for the given runtime (i.e. this method is hardcoded to return None). Only useful in an offchain context.

fn submit_vote_extrinsic(
&self,
__runtime_api_at_param__: &BlockId<Block>,
signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
) -> Result<(), ApiError>

Submit farmer vote vote that is essentially a header with bigger solution range than acceptable for block authoring. Only useful in an offchain context.

fn submit_vote_extrinsic_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
) -> Result<(), ApiError>

Submit farmer vote vote that is essentially a header with bigger solution range than acceptable for block authoring. Only useful in an offchain context.

fn is_in_block_list(
&self,
__runtime_api_at_param__: &BlockId<Block>,
farmer_public_key: &FarmerPublicKey
) -> Result<bool, ApiError>

Check if farmer_public_key is in block list (due to equivocation)

fn is_in_block_list_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
farmer_public_key: &FarmerPublicKey
) -> Result<bool, ApiError>

Check if farmer_public_key is in block list (due to equivocation)

fn total_pieces(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<NonZeroU64, ApiError>

Total number of pieces in a blockchain

fn total_pieces_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<NonZeroU64, ApiError>

Total number of pieces in a blockchain

fn records_root(
&self,
__runtime_api_at_param__: &BlockId<Block>,
segment_index: SegmentIndex
) -> Result<Option<RecordsRoot>, ApiError>

Get the merkle tree root of records for specified segment index

fn records_root_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
segment_index: SegmentIndex
) -> Result<Option<RecordsRoot>, ApiError>

Get the merkle tree root of records for specified segment index

fn extract_root_blocks(
&self,
__runtime_api_at_param__: &BlockId<Block>,
ext: &Block::Extrinsic
) -> Result<Option<Vec<RootBlock>>, ApiError>

Returns Vec<RootBlock> if a given extrinsic has them.

fn extract_root_blocks_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
ext: &Block::Extrinsic
) -> Result<Option<Vec<RootBlock>>, ApiError>

Returns Vec<RootBlock> if a given extrinsic has them.

fn root_plot_public_key(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<FarmerPublicKey>, ApiError>

Returns root plot public key in case block authoring is restricted.

fn root_plot_public_key_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<FarmerPublicKey>, ApiError>

Returns root plot public key in case block authoring is restricted.

fn should_adjust_solution_range(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<bool, ApiError>

Whether solution range adjustment is enabled.

fn should_adjust_solution_range_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<bool, ApiError>

Whether solution range adjustment is enabled.

fn chain_constants(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<ChainConstants, ApiError>

Get Subspace blockchain constants

fn chain_constants_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<ChainConstants, ApiError>

Get Subspace blockchain constants