Crate sp_consensus_subspace

Primitives for Subspace consensus.

Modules

digests

Private implementation details of Subspace consensus digests.

inherents

Inherents for Subspace consensus

offence

Common traits and types that are useful for describing offences for usage in environments that use subspace consensus.

Structs

GlobalRandomnesses

Subspace global randomnesses used for deriving global challenges.

Salts

Subspace salts used for challenges.

SolutionRanges

Subspace solution ranges used for challenges.

Traits

SubspaceApi

API necessary for block authorship with Subspace.

Functions

check_equivocation_proof

Verifies the equivocation proof by making sure that: both headers have different hashes, are targeting the same slot, and have valid signatures by the same authority.

Type Definitions

EquivocationProof

An equivocation proof for multiple block authorships on the same slot (i.e. double vote).

FarmerPublicKey

A Subspace farmer identifier. Necessarily equivalent to the schnorrkel public key used in the main Subspace module. If that ever changes, then this must, too.

FarmerSignature

A Subspace farmer signature.

SubspaceBlockWeight

The cumulative weight of a Subspace block, i.e. sum of block weights starting at this block until the genesis block.