Crate sp_consensus_subspace

Primitives for Subspace consensus.

Modules

consensus

Consensus-related runtime interface

digests

Private implementation details of Subspace consensus digests.

inherents

Inherents for Subspace consensus

offence

Common traits and types that are useful for describing offences for usage in environments that use subspace consensus.

Structs

KzgExtension

A KZG extension.

PosExtension

A Poof of space extension.

PotExtension

A Poof of time extension.

PotNextSlotInput

Next slot input for proof of time evaluation

PotParametersChange

Change of parameters to apply to PoT chain

SignedVote

Signed farmer vote.

SolutionRanges

Subspace solution ranges used for challenges.

WrappedPotOutput

Wrapped proof of time output for the purposes of runtime interface.

WrappedSolution

Wrapped solution for the purposes of runtime interface.

WrappedVerifySolutionParams

Wrapped solution verification parameters for the purposes of runtime interface.

Enums

ChainConstants

Subspace blockchain constants.

PotParameters

Proof of time parameters

SubspaceJustification

Subspace justification

Vote

Farmer vote.

Traits

SubspaceApi

API necessary for block authorship with Subspace.

Functions

is_equivocation_proof_valid

Verifies the equivocation proof by making sure that: both headers have different hashes, are targeting the same slot, and have valid signatures by the same authority.

Type Aliases

EquivocationProof

An equivocation proof for multiple block authorships on the same slot (i.e. double vote).

FarmerPublicKey

A Subspace farmer identifier. Necessarily equivalent to the schnorrkel public key used in the main Subspace module. If that ever changes, then this must, too.

FarmerSignature

A Subspace farmer signature.