sp_consensus_subspace

Module digests

Private implementation details of Subspace consensus digests.

Structs

SubspaceDigestItems

Digest items extracted from a header into convenient form

Enums

Error

Digest error

ErrorDigestType

Various kinds of digest types used in errors

PreDigest

A Subspace pre-runtime digest. This contains all data required to validate a block and for the Subspace runtime module.

PreDigestPotInfo

Proof of time information in pre-digest

Traits

CompatibleDigestItem

A digest item which is usable with Subspace consensus.

Functions

extract_pre_digest

Extract the Subspace pre digest from the given header. Pre-runtime digests are mandatory, the function will return Err if none is found.

extract_subspace_digest_items

Extract the Subspace global randomness from the given header.