Trait sp_consensus_subspace::digests::CompatibleDigestItem

pub trait CompatibleDigestItem: Sized {
    fn subspace_pre_digest<AccountId: Encode>(
        pre_digest: &PreDigest<AccountId>
    ) -> Self;
    fn as_subspace_pre_digest<AccountId: Decode>(
        &self
    ) -> Option<PreDigest<AccountId>>;
    fn subspace_seal(signature: FarmerSignature) -> Self;
    fn as_subspace_seal(&self) -> Option<FarmerSignature>;
    fn global_randomness_descriptor(
        global_randomness: GlobalRandomnessDescriptor
    ) -> Self;
    fn as_global_randomness_descriptor(
        &self
    ) -> Option<GlobalRandomnessDescriptor>;
    fn solution_range_descriptor(
        solution_range: SolutionRangeDescriptor
    ) -> Self;
    fn as_solution_range_descriptor(&self) -> Option<SolutionRangeDescriptor>;
    fn salt_descriptor(salt: SaltDescriptor) -> Self;
    fn as_salt_descriptor(&self) -> Option<SaltDescriptor>;
}

A digest item which is usable with Subspace consensus.

Required Methods

fn subspace_pre_digest<AccountId: Encode>(
pre_digest: &PreDigest<AccountId>
) -> Self

Construct a digest item which contains a Subspace pre-digest.

fn as_subspace_pre_digest<AccountId: Decode>(
&self
) -> Option<PreDigest<AccountId>>

If this item is an Subspace pre-digest, return it.

fn subspace_seal(signature: FarmerSignature) -> Self

Construct a digest item which contains a Subspace seal.

fn as_subspace_seal(&self) -> Option<FarmerSignature>

If this item is a Subspace signature, return the signature.

fn global_randomness_descriptor(
global_randomness: GlobalRandomnessDescriptor
) -> Self

Construct a digest item which contains a global randomness descriptor.

fn as_global_randomness_descriptor(&self) -> Option<GlobalRandomnessDescriptor>

If this item is a Subspace global randomness descriptor, return it.

fn solution_range_descriptor(solution_range: SolutionRangeDescriptor) -> Self

Construct a digest item which contains a solution range descriptor.

fn as_solution_range_descriptor(&self) -> Option<SolutionRangeDescriptor>

If this item is a Subspace solution range descriptor, return it.

fn salt_descriptor(salt: SaltDescriptor) -> Self

Construct a digest item which contains a salt descriptor.

fn as_salt_descriptor(&self) -> Option<SaltDescriptor>

If this item is a Subspace salt descriptor, return it.