Trait sp_consensus_subspace::digests::CompatibleDigestItem

pub trait CompatibleDigestItem: Sized {

    // Required methods
    fn subspace_pre_digest<AccountId: Encode>(
        pre_digest: &PreDigest<FarmerPublicKey, AccountId>
    ) -> Self;
    fn as_subspace_pre_digest<AccountId: Decode>(
        &self
    ) -> Option<PreDigest<FarmerPublicKey, AccountId>>;
    fn subspace_seal(signature: FarmerSignature) -> Self;
    fn as_subspace_seal(&self) -> Option<FarmerSignature>;
    fn global_randomness(global_randomness: Randomness) -> Self;
    fn as_global_randomness(&self) -> Option<Randomness>;
    fn solution_range(solution_range: SolutionRange) -> Self;
    fn as_solution_range(&self) -> Option<SolutionRange>;
    fn next_global_randomness(global_randomness: Randomness) -> Self;
    fn as_next_global_randomness(&self) -> Option<Randomness>;
    fn next_solution_range(solution_range: SolutionRange) -> Self;
    fn as_next_solution_range(&self) -> Option<SolutionRange>;
    fn segment_commitment(
        segment_index: SegmentIndex,
        segment_commitment: SegmentCommitment
    ) -> Self;
    fn as_segment_commitment(&self) -> Option<(SegmentIndex, SegmentCommitment)>;
    fn enable_solution_range_adjustment_and_override(
        override_solution_range: Option<SolutionRange>
    ) -> Self;
    fn as_enable_solution_range_adjustment_and_override(
        &self
    ) -> Option<Option<SolutionRange>>;
    fn root_plot_public_key_update(
        root_plot_public_key: Option<FarmerPublicKey>
    ) -> Self;
    fn as_root_plot_public_key_update(&self) -> Option<Option<FarmerPublicKey>>;

}

A digest item which is usable with Subspace consensus.

Required Methods

fn subspace_pre_digest<AccountId: Encode>( pre_digest: &PreDigest<FarmerPublicKey, AccountId> ) -> Self

Construct a digest item which contains a Subspace pre-digest.

fn as_subspace_pre_digest<AccountId: Decode>( &self ) -> Option<PreDigest<FarmerPublicKey, AccountId>>

If this item is an Subspace pre-digest, return it.

fn subspace_seal(signature: FarmerSignature) -> Self

Construct a digest item which contains a Subspace seal.

fn as_subspace_seal(&self) -> Option<FarmerSignature>

If this item is a Subspace signature, return the signature.

fn global_randomness(global_randomness: Randomness) -> Self

Construct a digest item which contains a global randomness.

fn as_global_randomness(&self) -> Option<Randomness>

If this item is a Subspace global randomness, return it.

fn solution_range(solution_range: SolutionRange) -> Self

Construct a digest item which contains a solution range.

fn as_solution_range(&self) -> Option<SolutionRange>

If this item is a Subspace solution range, return it.

fn next_global_randomness(global_randomness: Randomness) -> Self

Construct a digest item which contains next global randomness.

fn as_next_global_randomness(&self) -> Option<Randomness>

If this item is a Subspace next global randomness, return it.

fn next_solution_range(solution_range: SolutionRange) -> Self

Construct a digest item which contains next solution range.

fn as_next_solution_range(&self) -> Option<SolutionRange>

If this item is a Subspace next solution range, return it.

fn segment_commitment( segment_index: SegmentIndex, segment_commitment: SegmentCommitment ) -> Self

Construct a digest item which contains segment commitment.

fn as_segment_commitment(&self) -> Option<(SegmentIndex, SegmentCommitment)>

If this item is a Subspace segment commitment, return it.

fn enable_solution_range_adjustment_and_override( override_solution_range: Option<SolutionRange> ) -> Self

Construct digest item than indicates enabling of solution range adjustment and override next solution range.

fn as_enable_solution_range_adjustment_and_override( &self ) -> Option<Option<SolutionRange>>

If this item is a Subspace Enable solution range adjustment and override next solution range, return it.

fn root_plot_public_key_update( root_plot_public_key: Option<FarmerPublicKey> ) -> Self

Construct digest item than indicates update of root plot public key.

fn as_root_plot_public_key_update(&self) -> Option<Option<FarmerPublicKey>>

If this item is a Subspace update of root plot public key, return it.