Module sp_consensus_subspace::offence

Common traits and types that are useful for describing offences for usage in environments that use subspace consensus.

Comparison with sp_staking::offence

Unlike sp_staking::offence that handles both the offline and equivocation offences, there is only equivocation attack in subspace as it’s a permissionless consensus based on PoC holding to Nakamoto’s vision and does not have a known validator set for the block production.

Structs

OffenceDetails

A details about an offending authority for a particular kind of offence.

Enums

OffenceError

Errors that may happen on offence reports.

Traits

Offence

A trait implemented by an offence report.

OnOffenceHandler

A trait to take action on an offence.

ReportOffence

A trait for decoupling offence reporters from the actual handling of offence reports.

Type Aliases

Kind

The kind of an offence, is a byte string representing some kind identifier e.g. b"sub:equivocation"