Trait sp_consensus_subspace::offence::Offence

pub trait Offence<Offender> {
    type TimeSlot: Clone + Codec + Ord;

    const ID: Kind;

    fn offenders(&self) -> Vec<Offender> ;
    fn time_slot(&self) -> Self::TimeSlot;
}

A trait implemented by an offence report.

This trait assumes that the offence is legitimate and was validated already.

Examples of offences include: a BABE equivocation or a GRANDPA unjustified vote.

Required Associated Types

type TimeSlot: Clone + Codec + Ord

A type that represents a point in time on an abstract timescale.

See Offence::time_slot for details. The only requirement is that such timescale could be represented by a single u128 value.

Required Associated Constants

const ID: Kind

Identifier which is unique for this kind of an offence.

Required Methods

fn offenders(&self) -> Vec<Offender>

The list of all offenders involved in this incident.

The list has no duplicates, so it is rather a set.

fn time_slot(&self) -> Self::TimeSlot

A point in time when this offence happened.

This is used for looking up offences that happened at the “same time”.

The timescale is abstract and doesn’t have to be the same across different implementations of this trait. The value doesn’t represent absolute timescale though since it is interpreted along with the session_index. Two offences are considered to happen at the same time if time_slot is the same.