Trait sp_consensus_subspace::SubspaceApi

pub trait SubspaceApi<Block: BlockT, RewardAddress: Encode + Decode>: Core<Block> {

    fn confirmation_depth_k(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<<<Block as BlockT>::Header as HeaderT>::Number, ApiError> { ... }
    fn confirmation_depth_k_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<<<Block as BlockT>::Header as HeaderT>::Number, ApiError> { ... }
    fn record_size(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u32, ApiError> { ... }
    fn record_size_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u32, ApiError> { ... }
    fn recorded_history_segment_size(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u32, ApiError> { ... }
    fn recorded_history_segment_size_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u32, ApiError> { ... }
    fn max_plot_size(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, ApiError> { ... }
    fn max_plot_size_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, ApiError> { ... }
    fn slot_duration(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Duration, ApiError> { ... }
    fn slot_duration_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Duration, ApiError> { ... }
    fn global_randomnesses(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<GlobalRandomnesses, ApiError> { ... }
    fn global_randomnesses_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<GlobalRandomnesses, ApiError> { ... }
    fn solution_ranges(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<SolutionRanges, ApiError> { ... }
    fn solution_ranges_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<SolutionRanges, ApiError> { ... }
    fn salts(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Salts, ApiError> { ... }
    fn salts_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Salts, ApiError> { ... }
    fn submit_report_equivocation_extrinsic(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        equivocation_proof: EquivocationProof<Block::Header>
    ) -> Result<Option<()>, ApiError> { ... }
    fn submit_report_equivocation_extrinsic_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        equivocation_proof: EquivocationProof<Block::Header>
    ) -> Result<Option<()>, ApiError> { ... }
    fn submit_vote_extrinsic(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
    ) -> Result<(), ApiError> { ... }
    fn submit_vote_extrinsic_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
    ) -> Result<(), ApiError> { ... }
    fn is_in_block_list(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        farmer_public_key: &FarmerPublicKey
    ) -> Result<bool, ApiError> { ... }
    fn is_in_block_list_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        farmer_public_key: &FarmerPublicKey
    ) -> Result<bool, ApiError> { ... }
    fn total_pieces(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, ApiError> { ... }
    fn total_pieces_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, ApiError> { ... }
    fn records_root(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        segment_index: u64
    ) -> Result<Option<Sha256Hash>, ApiError> { ... }
    fn records_root_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        segment_index: u64
    ) -> Result<Option<Sha256Hash>, ApiError> { ... }
    fn extract_root_blocks(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        ext: &Block::Extrinsic
    ) -> Result<Option<Vec<RootBlock>>, ApiError> { ... }
    fn extract_root_blocks_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        ext: &Block::Extrinsic
    ) -> Result<Option<Vec<RootBlock>>, ApiError> { ... }
    fn root_plot_public_key(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<FarmerPublicKey>, ApiError> { ... }
    fn root_plot_public_key_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<FarmerPublicKey>, ApiError> { ... }

}

API necessary for block authorship with Subspace.

Provided Methods

fn confirmation_depth_k(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<<<Block as BlockT>::Header as HeaderT>::Number, ApiError>

Depth K after which a block enters the recorded history (a global constant, as opposed to the client-dependent transaction confirmation depth k).

fn confirmation_depth_k_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<<<Block as BlockT>::Header as HeaderT>::Number, ApiError>

Depth K after which a block enters the recorded history (a global constant, as opposed to the client-dependent transaction confirmation depth k).

fn record_size(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u32, ApiError>

The size of data in one piece (in bytes).

fn record_size_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u32, ApiError>

The size of data in one piece (in bytes).

fn recorded_history_segment_size(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u32, ApiError>

Recorded history is encoded and plotted in segments of this size (in bytes).

fn recorded_history_segment_size_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u32, ApiError>

Recorded history is encoded and plotted in segments of this size (in bytes).

fn max_plot_size(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, ApiError>

Maximum number of pieces in each plot

fn max_plot_size_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, ApiError>

Maximum number of pieces in each plot

fn slot_duration(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Duration, ApiError>

The slot duration in milliseconds for Subspace.

fn slot_duration_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Duration, ApiError>

The slot duration in milliseconds for Subspace.

fn global_randomnesses(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<GlobalRandomnesses, ApiError>

Global randomnesses used for deriving global challenges.

fn global_randomnesses_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<GlobalRandomnesses, ApiError>

Global randomnesses used for deriving global challenges.

fn solution_ranges(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<SolutionRanges, ApiError>

Solution ranges.

fn solution_ranges_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<SolutionRanges, ApiError>

Solution ranges.

fn salts(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Salts, ApiError>

Subspace salts used for challenges.

fn salts_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Salts, ApiError>

Subspace salts used for challenges.

fn submit_report_equivocation_extrinsic(
&self,
__runtime_api_at_param__: &BlockId<Block>,
equivocation_proof: EquivocationProof<Block::Header>
) -> Result<Option<()>, ApiError>

Submits an unsigned extrinsic to report an equivocation. The caller must provide the equivocation proof. The extrinsic will be unsigned and should only be accepted for local authorship (not to be broadcast to the network). This method returns None when creation of the extrinsic fails, e.g. if equivocation reporting is disabled for the given runtime (i.e. this method is hardcoded to return None). Only useful in an offchain context.

fn submit_report_equivocation_extrinsic_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
equivocation_proof: EquivocationProof<Block::Header>
) -> Result<Option<()>, ApiError>

Submits an unsigned extrinsic to report an equivocation. The caller must provide the equivocation proof. The extrinsic will be unsigned and should only be accepted for local authorship (not to be broadcast to the network). This method returns None when creation of the extrinsic fails, e.g. if equivocation reporting is disabled for the given runtime (i.e. this method is hardcoded to return None). Only useful in an offchain context.

fn submit_vote_extrinsic(
&self,
__runtime_api_at_param__: &BlockId<Block>,
signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
) -> Result<(), ApiError>

Submit farmer vote vote that is essentially a header with bigger solution range than acceptable for block authoring. Only useful in an offchain context.

fn submit_vote_extrinsic_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
signed_vote: SignedVote<<<Block as BlockT>::Header as HeaderT>::Number, Block::Hash, RewardAddress>
) -> Result<(), ApiError>

Submit farmer vote vote that is essentially a header with bigger solution range than acceptable for block authoring. Only useful in an offchain context.

fn is_in_block_list(
&self,
__runtime_api_at_param__: &BlockId<Block>,
farmer_public_key: &FarmerPublicKey
) -> Result<bool, ApiError>

Check if farmer_public_key is in block list (due to equivocation)

fn is_in_block_list_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
farmer_public_key: &FarmerPublicKey
) -> Result<bool, ApiError>

Check if farmer_public_key is in block list (due to equivocation)

fn total_pieces(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, ApiError>

Total number of pieces in a blockchain

fn total_pieces_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, ApiError>

Total number of pieces in a blockchain

fn records_root(
&self,
__runtime_api_at_param__: &BlockId<Block>,
segment_index: u64
) -> Result<Option<Sha256Hash>, ApiError>

Get the merkle tree root of records for specified segment index

fn records_root_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
segment_index: u64
) -> Result<Option<Sha256Hash>, ApiError>

Get the merkle tree root of records for specified segment index

fn extract_root_blocks(
&self,
__runtime_api_at_param__: &BlockId<Block>,
ext: &Block::Extrinsic
) -> Result<Option<Vec<RootBlock>>, ApiError>

Returns Vec<RootBlock> if a given extrinsic has them.

fn extract_root_blocks_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
ext: &Block::Extrinsic
) -> Result<Option<Vec<RootBlock>>, ApiError>

Returns Vec<RootBlock> if a given extrinsic has them.

fn root_plot_public_key(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<FarmerPublicKey>, ApiError>

Returns root plot public key in case block authoring is restricted.

fn root_plot_public_key_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<FarmerPublicKey>, ApiError>

Returns root plot public key in case block authoring is restricted.